requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
google-cloud-vision==3.4.4
openai==1.3.5
gspread==5.12.0
oauth2client==4.1.3
Pillow==10.0.1
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.25.2
requests==2.31.0
pdfplumber==0.9.0
flask-cors==4.0.0
gunicorn==21.2.0 